% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dealpha}
\alias{dealpha}
\title{Remove alpha channel from hexadecimal color value}
\usage{
dealpha(x)
}
\arguments{
\item{x}{is a character vector of colors}
}
\value{
a character vector of colors with the alpha channel set to 'FF' (= 100%)
}
\description{
Remove alpha channel from hexadecimal color value
}
\examples{
dealpha(rainbow(5, alpha = 0.2))
}
