% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{Retrieve simple or rich color names}
\usage{
get_names(cols, table = c("plain", "rich"), maxColorValue = 255)
}
\arguments{
\item{cols}{a character vector of hexadecimal color codes, or (pending) a
matrix with red, green, blue values as columns.}

\item{table}{lookup table complexity, "plain" contains 228 unique color names whereas
"rich" calls an API that serves 17,375 color names.}
}
\value{
A character vector containing the names of colors in 'col'
}
\description{
Retrieve simple or rich color names
}
\details{
When the 'rich' table is specified, the function calls the Color Names API
    (https://github.com/meodai/color-names), which currently contains ~0.10\% of all
    >16.7 million RGB colors. The closest color name is returned. Users may submit
    candidate color names at the URL above.
}
\examples{
get_names(cols = "#ff23ef", table = "plain")
get_names(cols = "#ff23ef", table = "rich")

get_names(cols = c("#123456", "#ffffff"), table = "rich")
get_names(cols = rainbow(5), table = "plain")
}
