% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hex2Lab}
\alias{hex2Lab}
\title{Convert hexadecimal color strings to CIE Lab coordinates}
\usage{
hex2Lab(x, toRGB = list(gamma = FALSE), toXYZ = list(illuminant = "D65",
  observer = 2, RefWhite = get("XYZperfectreflectingdiffuser", envir =
  environment()), RGBModel = "sRGB", RefWhiteRGB = get("whitepointsRGB", envir =
  environment()), gamma = NA, RefWhiteIllum =
  get("XYZperfectreflectingdiffuser", envir = environment()), CAT = "Bradford",
  CATarray = get("ChromaticAdaptation", envir = environment())),
  toLab = list(illuminant = "D65", observer = 2, RefWhite =
  get("XYZperfectreflectingdiffuser", envir = environment())))
}
\arguments{
\item{x}{a character vector of hexadecimal color strings.}

\item{toRGB}{parameter options for intermediate conversion to sRGB. See
?colorspace::hex2RGB for details.}

\item{toXYZ}{parameter options for intermediate conversion to XYZ. See
?colorscience::RGB2XYZ for details.}

\item{toLab}{parameter options for intermediate conversion to CIE Lab. See
?colorscience::XYZ2Lab for details.}
}
\value{
a matrix of CIE Lab coordinates
}
\description{
Convert hexadecimal color strings to CIE Lab coordinates
}
\examples{
hex2Lab(x = c("#ffffff", "#59ceff"))
}
