% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.hexColor}
\alias{is.hexColor}
\title{Test whether vector values are valid hexadecimal color value}
\usage{
is.hexColor(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a logical vector
}
\description{
Test whether vector values are valid hexadecimal color value
}
\examples{
is.hexadecimal(c("#ffffff", "#123456", "#123", "#zabcde"))
}
