% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pal_dE00}
\alias{pal_dE00}
\title{Calculate pairwise delta E2000 distances for a color palette}
\usage{
pal_dE00(pal)
}
\arguments{
\item{pal}{a character vector containing hexadecimal color values}
}
\value{
a numeric matrix of pairwise delta E2000 color distances
}
\description{
Calculate pairwise delta E2000 distances for a color palette
}
\examples{
pal_dE00(c("#ffffff", "#000000", "#123456"))
}
